pyproject.toml
[tool.poetry]
name = "referral-system"
version = "0.1.0"
description = "A FastAPI referral system with PostgreSQL and Redis"
authors = ["Azeasy <azeasy@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.27"
pydantic = {extras = ["email"], version = "^2.10.6"}
sqlalchemy = "^2.0"
asyncpg = "^0.29"
alembic = "^1.13"
redis = "^5.0"
python-jose = "^3.3"
passlib = "^1.7"
bcrypt = "^4.0"
pydantic-settings = "^2.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
